+++ changelog.md
@@ -4,6 +4,7 @@
 [Ripley]
 * Add some null checks to try to catch issue with SelectDispel.
 * Add logging to Spell.Init
+* Add warning to players about potential rollback issues when SaveBiota fails.
 
 ### 2019-03-31
 [Ripley]